Added code for initial endpoints
