+++ Makefile
@@ -1,0 +1,5 @@
+all:
+	mvn clean package -Pwildfly-swarm
+
+run:
+	java -jar target/agr_api-swarm.jar
