+++ pom.xml
@@ -1,28 +1,37 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Copyright (C) 2016, 2017 Antonio Goncalves and others. Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License. -->
 <project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0">
-	
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>org.eclipse.microprofile.sample</groupId>
 	<version>1.0.0-SNAPSHOT</version>
 	<artifactId>agr_api</artifactId>
 	<packaging>war</packaging>
 	<name>AGR :: API</name>
-
 	<properties>
 		<skipLicenseCheck>true</skipLicenseCheck>
 		<maven.compiler.source>1.8</maven.compiler.source>
 		<maven.compiler.target>1.8</maven.compiler.target>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 	</properties>
+	
+	  <dependencyManagement>
+    <dependencies>
+      <dependency>
+        <groupId>org.wildfly.swarm</groupId>
+        <artifactId>bom</artifactId>
+        <version>2016.12.0</version>
+        <type>pom</type>
+        <scope>import</scope>
+      </dependency>
+    </dependencies>
+  </dependencyManagement>
+	
+	
 	<dependencies>
 		<!-- JBoss Logging required due to older version from resteasy client -->
 		<dependency>
 			<groupId>org.wildfly.swarm</groupId>
-			<artifactId>bom-all</artifactId>
-			<version>2016.12.0</version>
-			<type>pom</type>
-			<scope>provided</scope>
+			<artifactId>jaxrs</artifactId>
 		</dependency>
 		<dependency>
 			<groupId>javax</groupId>
@@ -59,13 +68,11 @@
 			<version>3.0.18.Final</version>
 			<scope>test</scope>
 		</dependency>
-
 	</dependencies>
 	<build>
-		<finalName>microprofile-sample-swagger</finalName>
+		<finalName>agr_api</finalName>
 		<pluginManagement>
 			<plugins>
-
 				<plugin>
 					<groupId>org.apache.maven.plugins</groupId>
 					<artifactId>maven-compiler-plugin</artifactId>
@@ -75,9 +82,6 @@
 						<compilerArgument>-XDignore.symbol.file</compilerArgument>
 					</configuration>
 				</plugin>
-
-
-
 				<plugin>
 					<groupId>org.apache.maven.plugins</groupId>
 					<artifactId>maven-jar-plugin</artifactId>
@@ -100,7 +104,6 @@
 						</archive>
 					</configuration>
 				</plugin>
-
 				<plugin>
 					<groupId>org.apache.maven.plugins</groupId>
 					<artifactId>maven-war-plugin</artifactId>
@@ -134,10 +137,10 @@
 							<swarm.http.port>8080</swarm.http.port>
 							<swarm.context.path>/api</swarm.context.path>
 						</properties>
+						<useUberJar>true</useUberJar>
 					</configuration>
 					<executions>
 						<execution>
-							<id>package</id>
 							<goals>
 								<goal>package</goal>
 							</goals>
@@ -155,6 +158,13 @@
 					<plugin>
 						<groupId>org.wildfly.swarm</groupId>
 						<artifactId>wildfly-swarm-plugin</artifactId>
+						<executions>
+							<execution>
+								<goals>
+									<goal>package</goal>
+								</goals>
+							</execution>
+						</executions>
 					</plugin>
 					<plugin>
 						<groupId>org.apache.maven.plugins</groupId>
