--- src/main/java/org/alliancegenome/api/controller/TopCDsEndpoint.java
@@ -1,62 +1,0 @@
-/*
- * Copyright (C) 2016, 2017 Antonio Goncalves and others.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
- * implied.
- *
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.alliancegenome.api.controller;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Random;
-
-import javax.enterprise.context.RequestScoped;
-import javax.json.Json;
-import javax.json.JsonArrayBuilder;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-
-import io.swagger.annotations.Api;
-
-@Path("/topcds")
-@Api(description = "Top CDs Endpoint")
-@RequestScoped
-public class TopCDsEndpoint {
-
-    @GET
-    @Produces(MediaType.APPLICATION_JSON)
-    public String getTopCDs() {
-
-        JsonArrayBuilder array = Json.createArrayBuilder();
-        List<Integer> randomCDs = getRandomNumbers();
-        for (Integer randomCD : randomCDs) {
-            array.add(Json.createObjectBuilder().add("id", randomCD));
-        }
-        return array.build().toString();
-    }
-
-    private List<Integer> getRandomNumbers() {
-        List<Integer> randomCDs = new ArrayList<>();
-        Random r = new Random();
-        randomCDs.add(r.nextInt(100) + 1101);
-        randomCDs.add(r.nextInt(100) + 1101);
-        randomCDs.add(r.nextInt(100) + 1101);
-        randomCDs.add(r.nextInt(100) + 1101);
-        randomCDs.add(r.nextInt(100) + 1101);
-
-        return randomCDs;
-    }
-}
