+++ src/main/java/org/alliancegenome/api/rest/GeneRESTInterface.java
@@ -1,0 +1,5 @@
+package org.alliancegenome.api.rest;
+
+public interface GeneRESTInterface {
+
+}
