+++ src/main/java/org/alliancegenome/api/service/GeneService.java
@@ -1,0 +1,8 @@
+package org.alliancegenome.api.service;
+
+import javax.enterprise.context.RequestScoped;
+
+@RequestScoped
+public class GeneService {
+
+}
