+++ src/main/java/org/alliancegenome/api/service/MessageService.java
@@ -1,0 +1,13 @@
+package org.alliancegenome.api.service;
+
+import javax.enterprise.context.RequestScoped;
+
+@RequestScoped
+public class MessageService {
+
+	public void message() {
+		System.out.println("Message Service:");
+		
+	}
+
+}
